chore: rename folder and create new withdraw request emails for state/CMS
