+++ lib/libs/email/content/index.ts
@@ -2,6 +2,6 @@ export * from "./new-submission";
 export * from "./tempExtension";
 export * from "./respondToRai";
 export * from "./widthdrawRai";
-export * from "./widthdrawPackage";
+export * from "./withdrawPackage";
 export * from "./email-components";
 export * from "./respondToRai";
