+++ lib/libs/email/content/withdrawPackage/emailTemplates/MedSpaCMS.tsx
