+++ lib/libs/email/content/withdrawPackage/emailTemplates/MedSpaState.tsx
